README1.md
This is me. Literally me. No other character can come close to relating to me like this. There is no way you can convince me this is not me. This character could not possibly be anymore me. It's me, and nobody can convince me otherwise. If anyone approached me on the topic of this not possibly being me, then I immediately shut them down with overwhelming evidence that this character is me. This character is me, it is indisputable. Why anyone would try to argue that this character is not me is beyond me. If you held two pictures of me and this character side by side, you'd see no difference. I can safely look at this character every day and say "Yup, that's me". I can practically see this character every time I look at myself in the mirror. I go outside and people stop me to comment how similar I look and act to this character. I chuckle softly as I'm assured everyday this character is me in every way. I can smile each time I get out of bed every morning knowing that I've found my identity with this character and I know my place in this world. It's really quite funny how similar this character is to me, it's almost like we're identical twins. When I first saw this character, I had an existential crisis. What if this character was the real me and I was the fictional being. What if this character actual became aware of my existence? Did this character have the ability to become self aware itself?